utils.run: take an array instead of a string
Also take an optional table of environ variables which will be *added*
to the child processes environment (so no need to include variables like
HOME etc).
run now takes an array - think of execvpe(2).

run_shell exists with the original functionality - running a string in a
shell, mainly for easy pipes; but ideally this won't be needed in the
future.